allow data from GraphDataGenerator to be used by graphics without formatter
